+++ RoungOut/core/src/Tests/BallTest.java
@@ -12,13 +12,13 @@ import static org.junit.jupiter.api.Assertions.*;
  */
 class BallTest {
 
-    private final static double THRESHOLD = 0.0001f;
+    private final static float THRESHOLD = 0.0001f;
 
-    private final static double XPOS = 100.0f;
-    private final static double YPOS = -100.0f;
-    private final static double RADIUS = 10f;
-    private final static double ANGLE = 4f*Math.PI;
-    private final static double SPEED = 20.0f;
+    private final static float XPOS = 100.0f;
+    private final static float YPOS = -100.0f;
+    private final static float RADIUS = 10f;
+    private final static float ANGLE = (float) (4f*Math.PI);
+    private final static float SPEED = 20.0f;
 
     private Ball ball;
 
@@ -50,22 +50,22 @@ class BallTest {
 
     @Test
     void setX() {
-        double expectedX = XPOS + 1f;
+        float expectedX = XPOS + 1f;
         ball.setX(expectedX);
         Assertions.assertEquals(expectedX, ball.getX(), THRESHOLD);
     }
 
     @Test
     void setY() {
-        double expectedY = YPOS - 1f;
+        float expectedY = YPOS - 1f;
         ball.setY(expectedY);
         Assertions.assertEquals(expectedY, ball.getY(), THRESHOLD);
     }
 
     @Test
     void setPosition() {
-        double expectedX = XPOS + 1f;
-        double expectedY = YPOS - 1f;
+        float expectedX = XPOS + 1f;
+        float expectedY = YPOS - 1f;
         ball.setPosition(expectedX, expectedY);
         Assertions.assertEquals(expectedX, ball.getX(), THRESHOLD);
         Assertions.assertEquals(expectedY, ball.getY(), THRESHOLD);
@@ -73,10 +73,10 @@ class BallTest {
 
     @Test
     void setAngle() {
-        double maxAngle = 4f*Math.PI;
-        double minAngle = -maxAngle;
-        for (double a = minAngle; a < maxAngle; a += maxAngle/64f) {
-            double expectedAngle = (a + 8f*Math.PI) % (2f*Math.PI);
+        float maxAngle = (float) (4f*Math.PI);
+        float minAngle = -maxAngle;
+        for (float a = minAngle; a < maxAngle; a += maxAngle/64f) {
+            float expectedAngle = (float) ((a + 8f*Math.PI) % (2f*Math.PI));
             ball.setAngle(a);
             Assertions.assertEquals(expectedAngle, ball.getAngle(), THRESHOLD);
         }
@@ -84,14 +84,14 @@ class BallTest {
 
     @Test
     void setSpeed() {
-        double expectedSpeed = SPEED + 100f;
+        float expectedSpeed = SPEED + 100f;
         ball.setSpeed(expectedSpeed);
         Assertions.assertEquals(expectedSpeed, ball.getSpeed(), THRESHOLD);
     }
 
     @Test
     void setMaxSpeed() {
-        double expectedSpeed = SPEED;
+        float expectedSpeed = SPEED;
         ball.setMaxSpeed(expectedSpeed);
         ball.setSpeed(expectedSpeed + 100f);
         Assertions.assertEquals(expectedSpeed, ball.getSpeed(), THRESHOLD);
